Rewritten NPC Scripts to be independant of NPCObjects Added CALL command to NPCs Removed TalkingMonster and RunawayMonster AI
