+++ ClientPackets.cs
@@ -755,21 +755,6 @@ namespace ClientPackets
             writer.Write(Key);
         }
     }
-    public sealed class TalkMonsterNPC : Packet
-    {
-        public override short Index { get { return (short)ClientPacketIds.TalkMonsterNPC; } }
-
-        public uint ObjectID;
-
-        protected override void ReadPacket(BinaryReader reader)
-        {
-            ObjectID = reader.ReadUInt32();
-        }
-        protected override void WritePacket(BinaryWriter writer)
-        {
-            writer.Write(ObjectID);
-        }
-    }
     public sealed class BuyItem : Packet
     {
         public override short Index { get { return (short)ClientPacketIds.BuyItem; } }
