+++ Legend of Mir.sln
@@ -11,7 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		ClientPackets.cs = ClientPackets.cs
 		Common.cs = Common.cs
-		Notes.txt = Notes.txt
 		ServerPackets.cs = ServerPackets.cs
 	EndProjectSection
 EndProject
