+++ Notes.txt
@@ -2,8 +2,4 @@ In development features - may need revisiting / refactoring / finishing
 
 -NPC $INPUTSTR, method it sets page names and calls (would be better in ProcessPage? bool AwaitingInput at top)
 
-
--MONSTER AI 200, runaway mob
--MONSTER AI 201, Talking mob
-
 -Character deletion code after x days, or under x level after y days
