+++ Server.MirForms/SMain.cs
@@ -408,7 +408,7 @@ namespace Server
 
         private void reloadNPCsToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            Envir.ReloadNPCs();
+            Envir.ReloadNPCs(null);
         }
 
         private void reloadDropsToolStripMenuItem_Click(object sender, EventArgs e)
