+++ Server/MirDatabase/NPCInfo.cs
@@ -38,13 +38,10 @@ namespace Server.MirDatabase
         public int FlagNeeded = 0;
         public int Conquest;
 
-        public bool IsDefault, IsRobot;
-
         public List<int> CollectQuestIndexes = new List<int>();
         public List<int> FinishQuestIndexes = new List<int>();
         
-        public NPCInfo()
-        { }
+        public NPCInfo() { }
         public NPCInfo(BinaryReader reader)
         {
             if (Envir.LoadVersion > 33)
