+++ Server/MirObjects/Monsters/VampireSpider.cs
@@ -144,11 +144,10 @@ namespace Server.MirObjects.Monsters
         {
             int damage = GetAttackPower(MinDC, MaxDC);
             if (damage == 0) return;
-
-            int value = 0;
-
             Point target = Functions.PointMove(CurrentLocation, Direction, 1);
 
+
+            int value;
             if (target == Target.CurrentLocation)
             {
                 value = Target.Attacked(this, damage, DefenceType.MACAgility);
