+++ Server/MirObjects/NPC/NPCActions.cs
@@ -1,0 +1,103 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Server.MirObjects
+{
+    public class NPCActions
+    {
+        public ActionType Type;
+        public List<string> Params = new List<string>();
+
+        public NPCActions(ActionType action, params string[] p)
+        {
+            Type = action;
+
+            Params.AddRange(p);
+        }
+    }
+
+    public enum ActionType
+    {
+        Move,
+        InstanceMove,
+        GiveGold,
+        TakeGold,
+        GiveGuildGold,
+        TakeGuildGold,
+        GiveCredit,
+        TakeCredit,
+        GiveItem,
+        TakeItem,
+        GiveExp,
+        GivePet,
+        ClearPets,
+        AddNameList,
+        DelNameList,
+        ClearNameList,
+        GiveHP,
+        GiveMP,
+        ChangeLevel,
+        SetPkPoint,
+        ReducePkPoint,
+        IncreasePkPoint,
+        ChangeGender,
+        ChangeClass,
+        LocalMessage,
+        Goto,
+        Call,
+        GiveSkill,
+        RemoveSkill,
+        Set,
+        Param1,
+        Param2,
+        Param3,
+        Mongen,
+        TimeRecall,
+        TimeRecallGroup,
+        BreakTimeRecall,
+        MonClear,
+        GroupRecall,
+        GroupTeleport,
+        DelayGoto,
+        Mov,
+        Calc,
+        GiveBuff,
+        RemoveBuff,
+        AddToGuild,
+        RemoveFromGuild,
+        RefreshEffects,
+        ChangeHair,
+        CanGainExp,
+        ComposeMail,
+        AddMailItem,
+        AddMailGold,
+        SendMail,
+        GroupGoto,
+        EnterMap,
+        GivePearls,
+        TakePearls,
+        MakeWeddingRing,
+        ForceDivorce,
+        GlobalMessage,
+        LoadValue,
+        SaveValue,
+        RemovePet,
+        ConquestGuard,
+        ConquestGate,
+        ConquestWall,
+        ConquestSiege,
+        TakeConquestGold,
+        SetConquestRate,
+        StartConquest,
+        ScheduleConquest,
+        OpenGate,
+        CloseGate,
+        Break,
+        AddGuildNameList,
+        DelGuildNameList,
+        ClearGuildNameList,
+        OpenBrowser,
+        GetRandomText,
+    }
+}
